requirements.txt
cryptography
dash
discord.py
keyboard
numpy
plotly
PyAutoGUI
PyGetWindow
pyperclip
PyYAML
Requests
scheduler
scipy
selenium==4.18.1
seleniumbase
